pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fbuild"
readme = "README.md"
description = "Modern replacement for PlatformIO with URL-based platform/toolchain management and bug-free architecture"
requires-python = ">=3.10"
keywords = ["embedded", "arduino", "platformio", "compiler", "toolchain", "firmware", "microcontroller"]
license = { text = "BSD 3-Clause License" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "requests>=2.28.0",
    "tqdm>=4.64.0",
    "pyserial>=3.5",
    "esptool>=4.6.0",
    "psutil>=5.9.0",
]
# Change this with the version number bump.
version = "0.0.1"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.setuptools]
packages = ["fbuild", "lint.ruff_plugins"]
package-dir = {"fbuild" = "src/fbuild", "lint.ruff_plugins" = "lint_plugins/ruff"}

[tool.ruff]
line-length = 200
extend-exclude = ["**/.zap/**", "**/.fbuild/**", "**/.build/**"]

[tool.black]
line-length = 200
extend-exclude = '''
/(
    \.zap
  | \.fbuild
  | \.build
)/
'''

[tool.pylint."MESSAGES CONTROL"]
good-names = [
    "c",
    "i",
    "ok",
    "id",
    "e",
    "f"
]
disable = [
    "missing-function-docstring",
    "missing-module-docstring"
]

[tool.isort]
profile = "black"
skip_glob = [
    "*/.fbuild/**",
    "*/.build/**",
    "*/.zap/**",
    "*/build/**",
    "*/__pycache__/**",
    "*/.venv/**",
    "*/.git/**",
    "*/.pytest_cache/**",
    "*/.mypy_cache/**",
    "*/.ruff_cache/**",
    "*.egg-info/**",
]

[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

[tool.pyright]
# Basic type checking (current default)
typeCheckingMode = "basic"
reportMissingImports = "none"
extraPaths = ["src"]
# Exclude test directories from strict parameter type checking (pytest fixtures don't need types)
exclude = ["**/test_*.py", "**/tests/**", "**/.zap/**", "**/.fbuild/**", "**/.build/**"]

# Enable useful warnings that catch real bugs
reportUnusedImport = "warning"
reportUnusedVariable = "warning"
reportUnusedClass = "warning"
reportUnusedFunction = "warning"
reportUnnecessaryTypeIgnoreComment = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnnecessaryComparison = "warning"
reportUnnecessaryCast = "warning"
reportImplicitStringConcatenation = "warning"

# Enforce parameter and return type annotations (but only for missing annotations, not unknown types)
reportMissingParameterType = "error"

[project.scripts]
fbuild = "fbuild.cli:main"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselected by default, run with --full)",
]
# By default, skip integration tests
addopts = "-m 'not integration'"
testpaths = ["tests"]
[project.entry-points."flake8.extension"]
KBI = "lint.ruff_plugins.keyboard_interrupt_checker:KeyboardInterruptChecker"
